Code was not tested yet - expecting errors
Much more efficient code that will initialize the BT device
faster. The code will check the stored configuration for a
baud rate and try to communicate with the BT deivce. If the
communication fails, then baud rate guess will be initiated
to determin what is the correct baud rate of the device.